+++ ting.h
@@ -2,6 +2,10 @@
 
 Common header for Widora Ting LoRA UART test program
 
+KEY WORDS:
+TS --  Time stamp
+
+
 --------------------------------------------------*/
 
 #include     <stdio.h>      /*标准输入输出定义*/
@@ -14,133 +18,15 @@ Common header for Widora Ting LoRA UART test program
 #include     <errno.h>      /*错误号定义*/
 #include     <stdbool.h>
 #include     <string.h>
-/***@brief  设置串口通信速率
-*@param  fd     类型 int  打开串口的文件句柄
-*@param  speed  类型 int  串口速度
-*@return  void*/
 
-int speed_arr[] = {B115200, B38400, B19200, B9600, B4800, B2400, B1200, B300,
-	    B38400, B19200, B9600, B4800, B2400, B1200, B300, };
-int name_arr[] = {115200,38400,  19200,  9600,  4800,  2400,  1200,  300,
-	    38400,  19200,  9600, 4800, 2400, 1200,  300, };
-void set_speed(int fd, int speed)
-{
-  int   i;
-  int   status;
-  struct termios   Opt;
-  tcgetattr(fd, &Opt);
-  for ( i= 0;  i < sizeof(speed_arr) / sizeof(int);  i++)
-   {
-   	if  (speed == name_arr[i])
-   	{
-   	    tcflush(fd, TCIOFLUSH);
-    	cfsetispeed(&Opt, speed_arr[i]);
-    	cfsetospeed(&Opt, speed_arr[i]);
-    	status = tcsetattr(fd, TCSANOW, &Opt);
-    	if  (status != 0)
-            perror("tcsetattr fd1");
-     	return;
-     	}
-   tcflush(fd,TCIOFLUSH);
-   }
-}
-/**
-*@brief   设置串口数据位，停止位和效验位
-*@param  fd     类型  int  打开的串口文件句柄*
-*@param  databits 类型  int 数据位   取值 为 7 或者8*
-*@param  stopbits 类型  int 停止位   取值为 1 或者2*
-*@param  parity  类型  int  效验类型 取值为N,E,O,,S
-*/
-int set_Parity(int fd,int databits,int stopbits,int parity)
-{
-	struct termios options;
- if  ( tcgetattr( fd,&options)  !=  0)
-  {
-  	perror("SetupSerial 1");
-  	return(false);
-  }
-  options.c_cflag &= ~CSIZE;
-  switch (databits) /*设置数据位数*/
-  {
-  	case 7:
-  		options.c_cflag |= CS7;
-  		break;
-  	case 8:
-		options.c_cflag |= CS8;
-		break;
-	default:
-		fprintf(stderr,"Unsupported data size\n");
-		return (false);
-	}
-  switch (parity)
-  	{
-  	case 'n':
-	case 'N':
-		options.c_cflag &= ~PARENB;   /* Clear parity enable */
-		options.c_iflag &= ~INPCK;     /* Enable parity checking */
-		break;
-	case 'o':
-	case 'O':
-		options.c_cflag |= (PARODD | PARENB);  /* 设置为奇效验*/
-		options.c_iflag |= INPCK;             /* Disnable parity checking */
-		break;
-	case 'e':
-	case 'E':
-		options.c_cflag |= PARENB;     /* Enable parity */
-		options.c_cflag &= ~PARODD;   /* 转换为偶效验*/
-		options.c_iflag |= INPCK;       /* Disnable parity checking */
-		break;
-	case 'S':
-	case 's':  /*as no parity*/
-		options.c_cflag &= ~PARENB;
-		options.c_cflag &= ~CSTOPB;
-		break;
-	default:
-		fprintf(stderr,"Unsupported parity\n");
-		return (false);
-		}
-  /* 设置停止位*/
-  switch (stopbits)
-  	{
-  	case 1:
-  		options.c_cflag &= ~CSTOPB;
-		break;
-	case 2:
-		options.c_cflag |= CSTOPB;
-		break;
-	default:
-		fprintf(stderr,"Unsupported stop bits\n");
-		return (false);
-	}
-  /* Set input parity option */
-  if (parity != 'n')
-  		options.c_iflag |= INPCK;
-    options.c_cc[VTIME] = 150; // 15 seconds
-    options.c_cc[VMIN] = 0;
 
-  tcflush(fd,TCIFLUSH); /* Update the options and do it NOW */
-  if (tcsetattr(fd,TCSANOW,&options) != 0)
-  	{
-  		perror("SetupSerial 3");
-		return (false);
-	}
-  return (true);
- }
-/**
-*@breif 打开串口
-*/
-int OpenDev(char *Dev)
-{
-int	fd = open( Dev, O_RDWR | O_NOCTTY );         //| O_NOCTTY | O_NDELAY
-	if (-1 == fd)
-		{ /*设置数据位数*/
-			perror("Can't Open Serial Port");
-			return -1;
-		}
-	else
-	return fd;
+struct timeval g_tm;
+char g_pstr_time[30]; //time stamp string
 
-}
+//---buffers
+char g_strUserRxBuff[512]; //--string from Ting LoRa RX, like "LR,6666,40,xxxx,xx......"  xxxx--payload
+char g_strUserTxBuff[512]; //--string ready to send for Ting LoRa Tx, like "xxxxxx,xxxxx,xx,,...."  xxxx- payload
+char g_strAtBuff[64]; //--string for AT cmd/replay buff, such as "OK\r\n","-063\r\nOK\r\n" etc.
 
 /*----------------------------------------------------------------------------------------
   send command string to LORA and get feedback string
@@ -149,7 +35,7 @@ int	fd = open( Dev, O_RDWR | O_NOCTTY );         //| O_NOCTTY | O_NDELAY
   Window: '\r'-- move cursor to the head of current line  '\n' --- start a new line
   Linux: '\r'--same above  '\n'-- start a new line and get cursor to the head of the new line.
 ------------------------------------------------------------------------------------------------*/
-extern char buff[];
+//extern char buff[];
 extern int fd;
 void sendCMD(const char* strCMD,int ndelay)
 {
@@ -165,15 +51,15 @@ void sendCMD(const char* strCMD,int ndelay)
    {
 	if(read(fd,&ctmp,1)>0)
 	{
-		buff[nb]=ctmp;
+		g_strAtBuff[nb]=ctmp;
 		if( ctmp=='\n' && nb>1)// end of return string
 		{
-			buff[nb]='\0'; //--get rid of '\n'
+			g_strAtBuff[nb]='\0'; //--get rid of '\n'
 			break;
 		}
 		else if(ctmp=='\r')
 		{
-			buff[nb]='\0'; // get rid of '\r'
+			g_strAtBuff[nb]='\0'; // get rid of '\r'
 		}
 		else if(ctmp=='\n' && nb<2)
 		{
@@ -187,10 +73,9 @@ void sendCMD(const char* strCMD,int ndelay)
     //buff[nread]='\0';
     strncpy(strtmp,strCMD,len);
     strtmp[len-2]='\0'; //--to  skip \r\n
-    printf("%s: %s\n",strtmp,buff);
+    printf("%s: %s\n",strtmp,g_strAtBuff);
 }
 
-//========================== TING DATA PARSE ===========================
 
 /*------------------------------------------------
 Compare two string to ascertain they are identical 
@@ -252,3 +137,11 @@ void parseTingLoraWordsArray(char* pstrTingLoraItems[])
 
 }
 
+/*----- return time stamp as char*----------*/
+void pstrGetTimeStamp(char *g_pstr_time)
+{
+     gettimeofday(&g_tm,NULL);
+     sprintf(g_pstr_time,"%ld.%ld", g_tm.tv_sec,g_tm.tv_usec);
+//     printf("Time Stamp:%s\n",pstr_time);
+//     return str_tm;
+}
